Product Rate datatype changed

+++ src/Core/Application/Cash/GetCashBalanceRequest.cs
@@ -1,16 +1,16 @@
 ﻿    namespace RewardsPlus.Application.Cash;
 
-public class GetCashBalanceRequest : IRequest<double>
+public class GetCashBalanceRequest : IRequest<decimal>
 {
 }
 
-public class GetCashBalanceRequestHandler : IRequestHandler<GetCashBalanceRequest, double>
+public class GetCashBalanceRequestHandler : IRequestHandler<GetCashBalanceRequest, decimal>
 {
     private readonly ICashierService _cashierService;
 
     public GetCashBalanceRequestHandler(ICashierService cashierService) => _cashierService = cashierService;
 
-    public async Task<double> Handle(GetCashBalanceRequest request, CancellationToken cancellationToken) =>
+    public async Task<decimal> Handle(GetCashBalanceRequest request, CancellationToken cancellationToken) =>
                                                         await _cashierService.GetBalanceAsync();
 
 }
