+++ src/Core/Application/Payment/PayRequest.cs
@@ -2,8 +2,8 @@
 public class PayRequest
 {
     public string UserName { get; set; }
-    public double Amount { get; set; }
-    public PayRequest(string userName, double amount)
+    public decimal Amount { get; set; }
+    public PayRequest(string userName, decimal amount)
     {
         UserName = userName;
         Amount = amount;
